requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
textblob==0.17.1
aiohttp==3.8.5
requests==2.31.0
tweepy==4.14.0
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-asyncio==0.21.1

